+++ src/Asymmetric/Config.php
@@ -23,10 +23,6 @@ use ParagonIE\Halite\{
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- *
- * @property string|bool ENCODING
- * @property string HASH_DOMAIN_SEPARATION
- * @property bool HASH_SCALARMULT
  */
 final class Config extends BaseConfig
 {