+++ src/Config.php
@@ -21,10 +21,16 @@ use function array_key_exists;
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  *
+ * @property string|bool ENCODING
+ *
+ * AsymmetricCrypto:
+ * @property string HASH_DOMAIN_SEPARATION
+ * @property bool HASH_SCALARMULT
+ *
+ * SymmetricCrypto:
  * @property bool CHECKSUM_PUBKEY
  * @property int BUFFER
  * @property int HASH_LEN
- * @property string|bool ENCODING
  * @property int SHORTEST_CIPHERTEXT_LENGTH
  * @property int NONCE_BYTES
  * @property int HKDF_SALT_LEN
@@ -32,6 +38,7 @@ use function array_key_exists;
  * @property string MAC_ALGO
  * @property int MAC_SIZE
  * @property int PUBLICKEY_BYTES
+ * @property bool HKDF_USE_INFO
  * @property string HKDF_SBOX
  * @property string HKDF_AUTH
  * @property bool USE_PAE